Review fixes - Move the parser to definition to `discover.rb` - Move the rdoc plugin files to rdoc_plugin directory - Rename files - Rewrite signatures for the above changes
